Refactor, add test with file upload for blob
