+++ Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 anyhow = "1.0.86"
 # backtrace-on-stack-overflow = "0.3.0"
 bytes = "1.6.1"
+iroh-blobs = "0.24.0"
 # futures-lite = "2.3.0"
 # iroh-blobs = "0.23.0"
 # serde = "1.0.207"
